refactoring linking logic into fixture
also refactored the directory structure
